Refactor and support new Concourse version
Newer versions of Concourse return the Bearer token in the skymarshal_auth0
cookie instead of skymarshal_auth.

Adding list_teams and get_team.

We now allow working with a token only istead of login/password.

We should never concatenate urls with string formating because if a user gives
a trailing slash, with string formating it doesn't work. We now use
urllib.parse.urljoin which is designed to solve this issue.
Same issue with with path joining, the right way to go is by using os.path.join
instead.

We also add linting and code checking.